+++ src/Bank/core/Bank.java
@@ -1,4 +1,8 @@
-package MethodTesting;
+package Bank.core;
+import Bank.account.BankAccount;
+import Bank.account.CheckingAccount;
+import Bank.account.SavingsAccount;
+
 import java.util.HashMap;
 import java.util.ArrayList;
 import java.util.Random;
