+++ src/Bank/core/Main.java
@@ -1,4 +1,4 @@
-package MethodTesting;
+package Bank.core;
 
 public class Main {
     public static void main(String[] args) {
